Extract `ConstraintType` to separate module
