+++ cooper/constraints/__init__.py
@@ -1,3 +1,4 @@
 from .constraint import Constraint
-from .constraint_state import ConstraintMeasurement, ConstraintState, ConstraintType
+from .constraint_state import ConstraintMeasurement, ConstraintState
+from .constraint_type import ConstraintType
 from .slacks import ConstantSlack, DenseSlack, ExplicitSlack, IndexedSlack, SlackVariable
