+++ cooper/constraints/constraint_state.py
@@ -1,13 +1,7 @@
 from dataclasses import dataclass
-from enum import Enum, auto
 from typing import Optional
 
 import torch
-
-
-class ConstraintType(Enum):
-    EQUALITY = auto()
-    INEQUALITY = auto()
 
 
 @dataclass
