+++ cooper/constraints/constraint_type.py
@@ -1,0 +1,8 @@
+from enum import Enum, auto
+
+
+class ConstraintType(Enum):
+    """Constraint type enumeration."""
+
+    EQUALITY = auto()
+    INEQUALITY = auto()
